test : 내 정보 요청 인수 테스트 추가 보완
